Split traits.rs
