+++ src/prelude.rs
@@ -5,4 +5,6 @@ pub use hermite::HermiteMatrix;
 pub use triangular::*;
 pub use util::*;
 pub use assert::*;
-pub use traits::*;
+
+pub use qr::*;
+pub use opnorm::*;
